requirements.txt
pygame-ce>=2.20.0
moderngl>=5.8.0
PyGLM